+++ drivers/servo/servo_cluster.hpp
@@ -45,32 +45,32 @@ namespace servo {
     uint8_t count() const;
     uint8_t pin(uint8_t servo) const;
 
-    void enable(uint servo, bool load = true);
+    void enable(uint8_t servo, bool load = true);
     void enable(const uint8_t *servos, uint8_t length, bool load = true);
     void enable(std::initializer_list<uint8_t> servos, bool load = true);
     void enable_all(bool load = true);
 
-    void disable(uint servo, bool load = true);
+    void disable(uint8_t servo, bool load = true);
     void disable(const uint8_t *servos, uint8_t length, bool load = true);
     void disable(std::initializer_list<uint8_t> servos, bool load = true);
     void disable_all(bool load = true);
 
-    bool is_enabled(uint servo) const;
+    bool is_enabled(uint8_t servo) const;
 
-    float pulse(uint servo) const;
-    void pulse(uint servo, float pulse, bool load = true);
+    float pulse(uint8_t servo) const;
+    void pulse(uint8_t servo, float pulse, bool load = true);
     void pulse(const uint8_t *servos, uint8_t length, float pulse, bool load = true);
     void pulse(std::initializer_list<uint8_t> servos, float pulse, bool load = true);
     void all_to_pulse(float pulse, bool load = true);
 
-    float value(uint servo) const;
-    void value(uint servo, float value, bool load = true);
+    float value(uint8_t servo) const;
+    void value(uint8_t servo, float value, bool load = true);
     void value(const uint8_t *servos, uint8_t length, float value, bool load = true);
     void value(std::initializer_list<uint8_t> servos, float value, bool load = true);
     void all_to_value(float value, bool load = true);
 
-    float phase(uint servo) const;
-    void phase(uint servo, float phase, bool load = true);
+    float phase(uint8_t servo) const;
+    void phase(uint8_t servo, float phase, bool load = true);
     void phase(const uint8_t *servos, uint8_t length, float phase, bool load = true);
     void phase(std::initializer_list<uint8_t> servos, float phase, bool load = true);
     void all_to_phase(float phase, bool load = true);
@@ -79,43 +79,43 @@ namespace servo {
     bool frequency(float freq);
 
     //--------------------------------------------------
-    float min_value(uint servo) const;
-    float mid_value(uint servo) const;
-    float max_value(uint servo) const;
+    float min_value(uint8_t servo) const;
+    float mid_value(uint8_t servo) const;
+    float max_value(uint8_t servo) const;
 
-    void to_min(uint servo, bool load = true);
+    void to_min(uint8_t servo, bool load = true);
     void to_min(const uint8_t *servos, uint8_t length, bool load = true);
     void to_min(std::initializer_list<uint8_t> servos, bool load = true);
     void all_to_min(bool load = true);
 
-    void to_mid(uint servo, bool load = true);
+    void to_mid(uint8_t servo, bool load = true);
     void to_mid(const uint8_t *servos, uint8_t length, bool load = true);
     void to_mid(std::initializer_list<uint8_t> servos, bool load = true);
     void all_to_mid(bool load = true);
 
-    void to_max(uint servo, bool load = true);
+    void to_max(uint8_t servo, bool load = true);
     void to_max(const uint8_t *servos, uint8_t length, bool load = true);
     void to_max(std::initializer_list<uint8_t> servos, bool load = true);
     void all_to_max(bool load = true);
 
-    void to_percent(uint servo, float in, float in_min = ServoState::ZERO_PERCENT, float in_max = ServoState::ONEHUNDRED_PERCENT, bool load = true);
+    void to_percent(uint8_t servo, float in, float in_min = ServoState::ZERO_PERCENT, float in_max = ServoState::ONEHUNDRED_PERCENT, bool load = true);
     void to_percent(const uint8_t *servos, uint8_t length, float in, float in_min = ServoState::ZERO_PERCENT, float in_max = ServoState::ONEHUNDRED_PERCENT, bool load = true);
     void to_percent(std::initializer_list<uint8_t> servos, float in, float in_min = ServoState::ZERO_PERCENT, float in_max = ServoState::ONEHUNDRED_PERCENT, bool load = true);
     void all_to_percent(float in, float in_min = ServoState::ZERO_PERCENT, float in_max = ServoState::ONEHUNDRED_PERCENT, bool load = true);
 
-    void to_percent(uint servo, float in, float in_min, float in_max, float value_min, float value_max, bool load = true);
+    void to_percent(uint8_t servo, float in, float in_min, float in_max, float value_min, float value_max, bool load = true);
     void to_percent(const uint8_t *servos, uint8_t length, float in, float in_min, float in_max, float value_min, float value_max, bool load = true);
     void to_percent(std::initializer_list<uint8_t> servos, float in, float in_min, float in_max, float value_min, float value_max, bool load = true);
     void all_to_percent(float in, float in_min, float in_max, float value_min, float value_max, bool load = true);
 
-    Calibration& calibration(uint servo);
-    const Calibration& calibration(uint servo) const;
+    Calibration& calibration(uint8_t servo);
+    const Calibration& calibration(uint8_t servo) const;
 
     void load();
 
     //--------------------------------------------------
   private:
-    void apply_pulse(uint servo, float pulse, bool load);
+    void apply_pulse(uint8_t servo, float pulse, bool load);
     void create_servo_states(CalibrationType default_type, bool auto_phase);
     void create_servo_states(const Calibration& calibration, bool auto_phase);
   };
